tests: Fix broken compile for BasicBrokerTest and ExtendedBrokerTest on SAMD

+++ tests/BasicBrokerTest/BasicBrokerTest.ino
@@ -110,7 +110,7 @@ test(BasicBrokerTest, ZoneEraBroker) {
 test(BasicBrokerTest, ZoneInfoBroker) {
   basic::ZoneInfoBroker info(&kZoneAmerica_Los_Angeles);
   assertEqual(kZoneNameAmerica_Los_Angeles, info.name());
-  assertEqual(0xb7f7e8f2, info.zoneId());
+  assertEqual((uint32_t) 0xb7f7e8f2, info.zoneId());
   assertEqual(2000, info.startYear());
   assertEqual(2050, info.untilYear());
   assertEqual(1, info.numEras());
